backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pymysql==1.1.0
python-decouple==3.8
django-cors-headers==4.3.1
cryptography==41.0.7
